Revert "Move GPS files from edge firmware to libraries"
This reverts commit d053b8d8564ca49c71d027bdff282849b38e1faf.
