+++ src/gps/EdgePlatform.cpp
