+++ src/gps/EdgePlatform.h
