+++ src/gps/location_service.cpp
